+++ src/main/java/br/com/henrique/service/FilialService.java
@@ -11,6 +11,7 @@ import org.springframework.stereotype.Service;
 import br.com.henrique.model.Filial;
 import br.com.henrique.model.FilialPK;
 import br.com.henrique.repository.FilialRepository;
+import br.com.henrique.service.exception.ObjectFoundException;
 import br.com.henrique.service.exception.ObjectNotFoundException;
 
 @Service
@@ -31,35 +32,36 @@ public class FilialService {
         return repositFilial.findAll(pageable);
     }    
     
-    
     // Busca por Filial
-//    public Filial findById(Integer codigoEmpresa, Integer codigoFilial) {
-//        Filial filialBusca =  repositFilial.findById(filial.getFilialPK()).orElse(null);
-//        
-//        if (filialBusca == null) {
-//            throw new ObjectNotFoundException("Filial nao encontrada !");
-//        }
-//        return filialBusca;
-//    }
-    
     public Filial findById(FilialPK filialPK) {
-        Filial filialBusca2 = repositFilial.findById(filialPK).orElse(null);
-        
-        if (filialBusca2 == null) {
+        Filial filialBusca = repositFilial.findById(filialPK).orElse(null);
+        if (filialBusca == null) {
             throw new ObjectNotFoundException("Filial nao encontrada !");
         }
-        return filialBusca2;
+        return filialBusca;
     }
     
     // Inclui Filial
     public Filial addFilial(Filial filial) {
-            return repositFilial.save(filial);
+        Filial filialBuscaID = repositFilial.findById(filial.getFilialPK()).orElse(null);
+        if (filialBuscaID != null) {
+            throw new ObjectFoundException("Filial já encontrada !");
+        } 	
+        
+        Filial filialBuscaCNPJ= repositFilial.findByCNPJ(filial.getCnpj());
+        if (filialBuscaCNPJ != null) {
+            throw new ObjectFoundException("CNPJ informado já encontra-se cadastrado para outra Filial !");
+        } 	        
+        return repositFilial.save(filial);
     }
     
     // Atualiza Filial
     public void updateFilial(FilialPK filialPK, 
                              Filial filial) {
-        Filial filialAtualizado = this.findById(filialPK);
+        Filial filialAtualizado = this.findById(filial.getFilialPK());
+        if (filialAtualizado != null) {
+            throw new ObjectNotFoundException("Filial nao encontrada !");
+        } 	        
         
         filialAtualizado.setNome(filial.getNome());
         filialAtualizado.setCnpj(filial.getCnpj());
@@ -70,7 +72,10 @@ public class FilialService {
     
     // Exclusão de Filial
     public void deletaFilial(FilialPK filialPK) {
-        this.findById(filialPK);
+        Filial filialExcluir = this.findById(filialPK);
+        if (filialExcluir != null) {
+            throw new ObjectNotFoundException("Filial nao encontrada !");
+        } 	   
         
         repositFilial.deleteById(filialPK);
     }
